requirements_The_Critical_Thinkers.txt
markdown-it-py==3.0.0
Pillow==9.5.0
requests==2.30.0
streamlit==1.27.2
streamlit==1.29.0